Added Function increase_serial($serial)
This function tries it best to increase a date based serial number
as "YYYYmmddXX" where XX is an daily counter. With this it is possible to
reflect 99 Changes per day in the serial number. The function returns the 
unmodified Serial in case of an failure to avoid problems within Serialnumbering
and reloading ...